Move col tests over
